fix: resolve firebase permissions error and improve resume filtering based on searchterms
